requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
Pillow>=8.2.0
matplotlib>=3.4.2
opencv-python>=4.5.3
scikit-image>=0.18.2 